+++ ios/Classes/NativeVideoRenderView.swift
@@ -1,0 +1,17 @@
+import AVFoundation
+import UIKit
+
+final class NativeVideoRenderView: UIView {
+  override class var layerClass: AnyClass {
+    AVPlayerLayer.self
+  }
+
+  private var playerLayer: AVPlayerLayer {
+    layer as! AVPlayerLayer
+  }
+
+  func setPlayer(_ player: AVPlayer?) {
+    playerLayer.player = player
+    playerLayer.videoGravity = .resizeAspectFill
+  }
+}