+++ lib/native_reels_player.dart
@@ -2,3 +2,4 @@ export 'src/native_reels_player_api.dart';
 export 'src/video_controller.dart';
 export 'src/video_models.dart';
 export 'src/video_playback_state.dart';
+export 'src/native_video_view.dart';